refactor: rename folder from commands to mdules
