+++ src/main.js
@@ -1,5 +1,5 @@
 import FileManager from "./cli/fileManager.js";
-import ArgParser from "./modules/args.js";
+import ArgParser from "./helpers/args.js";
 
 const parser = new ArgParser();
 const userName = parser.getArg('username') || 'Anonimus';
