+++ src/modules/basic/cp.js
