+++ src/modules/compress/compress.js
