+++ src/modules/compress/compressCommands.js
