+++ src/modules/compress/decompress.js
