+++ src/modules/hash/hash.js
