+++ src/modules/hash/hashCommands.js
