+++ src/modules/os/EOL.js
