+++ src/modules/os/cpus.js
